+++ source/slang-capture-replay/capture/capture-manager.h
@@ -3,7 +3,7 @@
 
 #include <filesystem>
 #include "parameter-encoder.h"
-#include "capture-format.h"
+#include "../capture-format.h"
 
 namespace SlangCapture
 {
