+++ source/slang-capture-replay/capture/slang-composite-component-type.cpp
@@ -1,53 +1,55 @@
-#include "capture-utility.h"
-#include "slang-entrypoint.h"
+#include "../capture-utility.h"
+#include "slang-composite-component-type.h"
 
 namespace SlangCapture
 {
-    EntryPointCapture::EntryPointCapture(slang::IEntryPoint* entryPoint, CaptureManager* captureManager)
-        : m_actualEntryPoint(entryPoint),
+    CompositeComponentTypeCapture::CompositeComponentTypeCapture(
+            slang::IComponentType* componentType, CaptureManager* captureManager)
+        : m_actualCompositeComponentType(componentType),
           m_captureManager(captureManager)
     {
-        SLANG_CAPTURE_ASSERT(m_actualEntryPoint != nullptr);
+        SLANG_CAPTURE_ASSERT(m_actualCompositeComponentType != nullptr);
         SLANG_CAPTURE_ASSERT(m_captureManager != nullptr);
 
-        m_entryPointHandle = reinterpret_cast<uint64_t>(m_actualEntryPoint.get());
-        slangCaptureLog(LogLevel::Verbose, "%s: %p\n", __PRETTY_FUNCTION__, entryPoint);
+        m_compositeComponentHandle = reinterpret_cast<uint64_t>(m_actualCompositeComponentType.get());
+        slangCaptureLog(LogLevel::Verbose, "%s: %p\n", __PRETTY_FUNCTION__, componentType);
     }
 
-    EntryPointCapture::~EntryPointCapture()
+    CompositeComponentTypeCapture::~CompositeComponentTypeCapture()
     {
-        m_actualEntryPoint->release();
+        m_actualCompositeComponentType->release();
     }
 
-    ISlangUnknown* EntryPointCapture::getInterface(const Guid& guid)
+    ISlangUnknown* CompositeComponentTypeCapture::getInterface(const Guid& guid)
     {
-        if(guid == EntryPointCapture::getTypeGuid())
+        if (guid == IComponentType::getTypeGuid())
+        {
             return static_cast<ISlangUnknown*>(this);
-        else
-            return nullptr;
+        }
+        return nullptr;
     }
 
-    SLANG_NO_THROW slang::ISession* EntryPointCapture::getSession()
+    SLANG_NO_THROW slang::ISession* CompositeComponentTypeCapture::getSession()
     {
         slangCaptureLog(LogLevel::Verbose, "%s\n", __PRETTY_FUNCTION__);
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_getSession, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_getSession, m_compositeComponentHandle);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        slang::ISession* session = m_actualEntryPoint->getSession();
+        slang::ISession* res = m_actualCompositeComponentType->getSession();
 
         {
-            encoder->encodeAddress(session);
+            encoder->encodeAddress(res);
             m_captureManager->endMethodCaptureAppendOutput();
         }
 
-        return session;
+        return res;
     }
 
-    SLANG_NO_THROW slang::ProgramLayout* EntryPointCapture::getLayout(
+    SLANG_NO_THROW slang::ProgramLayout* CompositeComponentTypeCapture::getLayout(
         SlangInt    targetIndex,
         slang::IBlob**     outDiagnostics)
     {
@@ -55,12 +57,12 @@ namespace SlangCapture
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_getLayout, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_getLayout, m_compositeComponentHandle);
             encoder->encodeInt64(targetIndex);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        slang::ProgramLayout* programLayout = m_actualEntryPoint->getLayout(targetIndex, outDiagnostics);
+        slang::ProgramLayout* programLayout = m_actualCompositeComponentType->getLayout(targetIndex, outDiagnostics);
 
         {
             encoder->encodeAddress(*outDiagnostics);
@@ -71,15 +73,15 @@ namespace SlangCapture
         return programLayout;
     }
 
-    SLANG_NO_THROW SlangInt EntryPointCapture::getSpecializationParamCount()
+    SLANG_NO_THROW SlangInt CompositeComponentTypeCapture::getSpecializationParamCount()
     {
         // No need to capture this call as it is just a query.
         slangCaptureLog(LogLevel::Verbose, "%s\n", __PRETTY_FUNCTION__);
-        SlangInt res = m_actualEntryPoint->getSpecializationParamCount();
+        SlangInt res = m_actualCompositeComponentType->getSpecializationParamCount();
         return res;
     }
 
-    SLANG_NO_THROW SlangResult EntryPointCapture::getEntryPointCode(
+    SLANG_NO_THROW SlangResult CompositeComponentTypeCapture::getEntryPointCode(
         SlangInt    entryPointIndex,
         SlangInt    targetIndex,
         slang::IBlob**     outCode,
@@ -89,13 +91,13 @@ namespace SlangCapture
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_getEntryPointCode, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_getEntryPointCode, m_compositeComponentHandle);
             encoder->encodeInt64(entryPointIndex);
             encoder->encodeInt64(targetIndex);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        SlangResult res = m_actualEntryPoint->getEntryPointCode(entryPointIndex, targetIndex, outCode, outDiagnostics);
+        SlangResult res = m_actualCompositeComponentType->getEntryPointCode(entryPointIndex, targetIndex, outCode, outDiagnostics);
 
         {
             encoder->encodeAddress(*outCode);
@@ -106,7 +108,7 @@ namespace SlangCapture
         return res;
     }
 
-    SLANG_NO_THROW SlangResult EntryPointCapture::getTargetCode(
+    SLANG_NO_THROW SlangResult CompositeComponentTypeCapture::getTargetCode(
         SlangInt    targetIndex,
         slang::IBlob** outCode,
         slang::IBlob** outDiagnostics)
@@ -115,12 +117,12 @@ namespace SlangCapture
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_getTargetCode, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_getTargetCode, m_compositeComponentHandle);
             encoder->encodeInt64(targetIndex);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        SlangResult res = m_actualEntryPoint->getTargetCode(targetIndex, outCode, outDiagnostics);
+        SlangResult res = m_actualCompositeComponentType->getTargetCode(targetIndex, outCode, outDiagnostics);
 
         {
             encoder->encodeAddress(*outCode);
@@ -131,7 +133,7 @@ namespace SlangCapture
         return res;
     }
 
-    SLANG_NO_THROW SlangResult EntryPointCapture::getResultAsFileSystem(
+    SLANG_NO_THROW SlangResult CompositeComponentTypeCapture::getResultAsFileSystem(
         SlangInt    entryPointIndex,
         SlangInt    targetIndex,
         ISlangMutableFileSystem** outFileSystem)
@@ -140,13 +142,13 @@ namespace SlangCapture
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_getResultAsFileSystem, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_getResultAsFileSystem, m_compositeComponentHandle);
             encoder->encodeInt64(entryPointIndex);
             encoder->encodeInt64(targetIndex);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        SlangResult res = m_actualEntryPoint->getResultAsFileSystem(entryPointIndex, targetIndex, outFileSystem);
+        SlangResult res = m_actualCompositeComponentType->getResultAsFileSystem(entryPointIndex, targetIndex, outFileSystem);
 
         {
             encoder->encodeAddress(*outFileSystem);
@@ -156,7 +158,7 @@ namespace SlangCapture
         return res;
     }
 
-    SLANG_NO_THROW void EntryPointCapture::getEntryPointHash(
+    SLANG_NO_THROW void CompositeComponentTypeCapture::getEntryPointHash(
         SlangInt    entryPointIndex,
         SlangInt    targetIndex,
         slang::IBlob**     outHash)
@@ -165,13 +167,13 @@ namespace SlangCapture
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_getEntryPointHash, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_getEntryPointHash, m_compositeComponentHandle);
             encoder->encodeInt64(entryPointIndex);
             encoder->encodeInt64(targetIndex);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        m_actualEntryPoint->getEntryPointHash(entryPointIndex, targetIndex, outHash);
+        m_actualCompositeComponentType->getEntryPointHash(entryPointIndex, targetIndex, outHash);
 
         {
             encoder->encodeAddress(*outHash);
@@ -179,7 +181,7 @@ namespace SlangCapture
         }
     }
 
-    SLANG_NO_THROW SlangResult EntryPointCapture::specialize(
+    SLANG_NO_THROW SlangResult CompositeComponentTypeCapture::specialize(
         slang::SpecializationArg const*    specializationArgs,
         SlangInt                    specializationArgCount,
         slang::IComponentType**            outSpecializedComponentType,
@@ -189,13 +191,13 @@ namespace SlangCapture
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_specialize, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_specialize, m_compositeComponentHandle);
             encoder->encodeInt64(specializationArgCount);
             encoder->encodeStructArray(specializationArgs, specializationArgCount);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        SlangResult res = m_actualEntryPoint->specialize(specializationArgs, specializationArgCount, outSpecializedComponentType, outDiagnostics);
+        SlangResult res = m_actualCompositeComponentType->specialize(specializationArgs, specializationArgCount, outSpecializedComponentType, outDiagnostics);
 
         {
             encoder->encodeAddress(*outSpecializedComponentType);
@@ -206,7 +208,7 @@ namespace SlangCapture
         return res;
     }
 
-    SLANG_NO_THROW SlangResult EntryPointCapture::link(
+    SLANG_NO_THROW SlangResult CompositeComponentTypeCapture::link(
         slang::IComponentType**            outLinkedComponentType,
         ISlangBlob**                outDiagnostics)
     {
@@ -214,11 +216,11 @@ namespace SlangCapture
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_link, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_link, m_compositeComponentHandle);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        SlangResult res = m_actualEntryPoint->link(outLinkedComponentType, outDiagnostics);
+        SlangResult res = m_actualCompositeComponentType->link(outLinkedComponentType, outDiagnostics);
 
         {
             encoder->encodeAddress(*outLinkedComponentType);
@@ -229,7 +231,7 @@ namespace SlangCapture
         return res;
     }
 
-    SLANG_NO_THROW SlangResult EntryPointCapture::getEntryPointHostCallable(
+    SLANG_NO_THROW SlangResult CompositeComponentTypeCapture::getEntryPointHostCallable(
         int                     entryPointIndex,
         int                     targetIndex,
         ISlangSharedLibrary**   outSharedLibrary,
@@ -239,13 +241,13 @@ namespace SlangCapture
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_getEntryPointHostCallable, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_getEntryPointHostCallable, m_compositeComponentHandle);
             encoder->encodeInt32(entryPointIndex);
             encoder->encodeInt32(targetIndex);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        SlangResult res = m_actualEntryPoint->getEntryPointHostCallable(entryPointIndex, targetIndex, outSharedLibrary, outDiagnostics);
+        SlangResult res = m_actualCompositeComponentType->getEntryPointHostCallable(entryPointIndex, targetIndex, outSharedLibrary, outDiagnostics);
 
         {
             encoder->encodeAddress(*outSharedLibrary);
@@ -256,19 +258,19 @@ namespace SlangCapture
         return res;
     }
 
-    SLANG_NO_THROW SlangResult EntryPointCapture::renameEntryPoint(
+    SLANG_NO_THROW SlangResult CompositeComponentTypeCapture::renameEntryPoint(
         const char* newName, IComponentType** outEntryPoint)
     {
         slangCaptureLog(LogLevel::Verbose, "%s\n", __PRETTY_FUNCTION__);
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_renameEntryPoint, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_renameEntryPoint, m_compositeComponentHandle);
             encoder->encodeString(newName);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        SlangResult res = m_actualEntryPoint->renameEntryPoint(newName, outEntryPoint);
+        SlangResult res = m_actualCompositeComponentType->renameEntryPoint(newName, outEntryPoint);
 
         {
             encoder->encodeAddress(*outEntryPoint);
@@ -278,7 +280,7 @@ namespace SlangCapture
         return res;
     }
 
-    SLANG_NO_THROW SlangResult EntryPointCapture::linkWithOptions(
+    SLANG_NO_THROW SlangResult CompositeComponentTypeCapture::linkWithOptions(
         IComponentType** outLinkedComponentType,
         uint32_t compilerOptionEntryCount,
         slang::CompilerOptionEntry* compilerOptionEntries,
@@ -288,13 +290,13 @@ namespace SlangCapture
 
         ParameterEncoder* encoder {};
         {
-            encoder = m_captureManager->beginMethodCapture(ApiCallId::IEntryPoint_linkWithOptions, m_entryPointHandle);
+            encoder = m_captureManager->beginMethodCapture(ApiCallId::ICompositeComponentType_linkWithOptions, m_compositeComponentHandle);
             encoder->encodeUint32(compilerOptionEntryCount);
             encoder->encodeStructArray(compilerOptionEntries, compilerOptionEntryCount);
             encoder = m_captureManager->endMethodCapture();
         }
 
-        SlangResult res = m_actualEntryPoint->linkWithOptions(outLinkedComponentType, compilerOptionEntryCount, compilerOptionEntries, outDiagnostics);
+        SlangResult res = m_actualCompositeComponentType->linkWithOptions(outLinkedComponentType, compilerOptionEntryCount, compilerOptionEntries, outDiagnostics);
 
         {
             encoder->encodeAddress(*outLinkedComponentType);
@@ -304,10 +306,4 @@ namespace SlangCapture
 
         return res;
     }
-
-    SLANG_NO_THROW slang::FunctionReflection* EntryPointCapture::getFunctionReflection()
-    {
-        return m_actualEntryPoint->getFunctionReflection();
-    }
-
 }
