+++ source/slang-capture-replay/capture/slang-filesystem.h
@@ -1,9 +1,9 @@
 #ifndef SLANG_FILE_SYSTEM_H
 #define SLANG_FILE_SYSTEM_H
 
-#include "../core/slang-com-object.h"
 #include "slang-com-helper.h"
 #include "slang-com-ptr.h"
+#include "../../core/slang-com-object.h"
 #include "capture-manager.h"
 
 namespace SlangCapture
