+++ source/slang-capture-replay/capture/slang-module.h
@@ -4,8 +4,8 @@
 #include "slang-com-ptr.h"
 #include "slang.h"
 #include "slang-com-helper.h"
-#include "../core/slang-smart-pointer.h"
-#include "../slang/slang-compiler.h"
+#include "../../core/slang-smart-pointer.h"
+#include "../../slang/slang-compiler.h"
 #include "slang-entrypoint.h"
 #include "capture-manager.h"
