+++ source/slang-capture-replay/capture/slang-session.cpp
@@ -1,4 +1,4 @@
-#include "capture-utility.h"
+#include "../capture-utility.h"
 #include "slang-session.h"
 #include "slang-entrypoint.h"
 #include "slang-composite-component-type.h"
