+++ source/slang-capture-replay/capture/slang-type-conformance.cpp
@@ -1,4 +1,4 @@
-#include "capture-utility.h"
+#include "../capture-utility.h"
 #include "slang-type-conformance.h"
 
 namespace SlangCapture
