+++ source/slang-capture-replay/replay/slang-decoder.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 #include <unordered_map>
-#include "capture-format.h"
+#include "../capture-format.h"
 #include "decoder-consumer.h"
 
 namespace SlangCapture
