+++ tools/slang-replay/main.cpp
@@ -1,0 +1,10 @@
+#include <stdio.h>
+#include <replay/file-processor.h>
+#include <replay/decoder-consumer.h>
+
+int main(int argc, char *argv[])
+{
+    // TODO: This is just a place holder binary
+    SlangCapture::FileProcessor fileProcessor("input.capture");
+    return 0;
+}
